Rewrite the `Ticket workaround`
Using `#include` we can make some kind of Applicative functor (in the ML
sense) so the idea is to keep the same API in Dummy_ticket and
Tezos_ticket, and using `include` in chusai_ticket. After that, we can
make specialized function in the scope of chusai_ticket.

Should be removed when the ticket/storage bug from Ligo will be fixed.