[issue 37] naïve and simplistic storage for blocks (#38)
Adds a simple storage of blocks on layer 1. Blocks are referenced in a map, a second map registering child relations is maintained to allow suppression of orphan blocks in case of block removal following dispute resolution.
A lib is implemented, with a simple smart contract for testing, not intended to be originated.